requirements.txt
absl-py==0.7.1
argh==0.26.2
astetik==1.9.8
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chances==0.1.6
chardet==3.0.4
colorama==0.4.1
colorlover==0.3.0
cufflinks==0.14.6
cycler==0.10.0
Cython==0.29.6
dateparser==0.7.1
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
filelock==3.0.10
gast==0.2.2
geonamescache==1.0.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
importlib-metadata==0.18
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kerasplotlib==0.1.4
kiwisolver==1.0.1
llvmlite==0.28.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.7
numba==0.43.0
numpy==1.16.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pathtools==0.1.2
patsy==0.5.1
pbr==5.1.3
pickleshare==0.7.5
plotly==3.7.1
pluggy==0.12.0
pmdarima==1.1.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
prophet==0.1.1
protobuf==3.7.0
py==1.8.0
Pygments==2.3.1
pyparsing==2.3.1
pyramid-arima==0.9.0
pyrsistent==0.14.11
pytest==5.0.1
python-dateutil==2.8.0
python-vlc==3.0.6109
pytz==2018.9
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
regex==2019.3.12
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
statsmodels==0.9.0
talos==0.4.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toml==0.10.0
tornado==6.0.2
tox==3.9.0
tqdm==4.32.1
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.5.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrangle==0.6.2
zipp==0.5.2